Refactor type checks
